made the flask backend modular
