+++ flaskr/main.py
@@ -1,0 +1,7 @@
+
+from flaskr import create_app
+
+
+if __name__ == '__main__':
+    app = create_app()
+    app.run()
