+++ flaskr/notification.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+notification = Blueprint('notification', __name__)
