+++ flaskr/static/resources/scripts/mapScript.js
@@ -9,21 +9,21 @@ fetch(url).then(response => response.json()).then(json => {
     }
 });
 
-console.log('This is from mapScript');
+// console.log('This is from mapScript');
 
 
 
-console.log(Object.values(data));
-console.log(data[0][0]);
+// console.log(Object.values(data));
+// console.log(data[0][0]);
 
 
-const map = L.map('map').setView([49.88, -119.12], 13);
+const map = L.map('map').setView([49.88, -119.49], 13);
 
 
 
 L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', { attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors' }).addTo(map);
 
-var marker = L.marker([data[0][0], data[0][1]]).addTo(map);
+var marker = L.marker([49.88, -119.49]).addTo(map);
 
 
 // function fromMapScript() {
