+++ flaskr/templates/Search.html
@@ -10,46 +10,12 @@
   <link rel="stylesheet" href="{{ url_for('static', filename='resources/styles/mapStyles.css') }}">
   <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"
     integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo=" crossorigin=""></script>
-    <script src="{{ url_for('static', filename='resources/scripts/dataScript.js')}}" defer></script>
-  
-  
+  <script src="{{ url_for('static', filename='resources/scripts/dataScript.js')}}" defer></script>
+  <script src="{{ url_for('static', filename='resources/scripts/mapScript.js') }}" defer></script>
 </head>
 
 <body>
-<<<<<<< HEAD
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-    crossorigin="anonymous"></script>
-    <!-- Navbar Component -->
-  <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <div class="container-fluid">
-      <a class="navbar-brand" href="/landing">QuakeQuest</a>
-      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav"
-        aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <div class="collapse navbar-collapse" id="navbarNav">
-        <ul class="navbar-nav">
-          <li class="nav-item">
-            <a class="nav-link" aria-current="page" href="/landing">Home</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="/search">Search</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/about">About</a>
-          </li>
-          <!-- needs to be dynamically updated if logged in-->
-          <li class="nav-item">
-            <a class="nav-link" href="/login">Sign In</a>
-          </li>
-        </ul>
-      </div>
-    </div>
-  </nav>
-=======
   {% include 'navbar.html' %}
->>>>>>> refs/remotes/origin/main
 
   <!-- Grid container-->
   <div class="container-fluid vh-95">
@@ -195,29 +161,8 @@
       </div>
     </div>
   </div>
-<<<<<<< HEAD
-
-  <!-- Footer Component-->
-  <div class="container">
-    <footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
-      <p class="col-md-4 mb-0 text-body-secondary">&copy; 2024 QuakeQuest</p>
-      <a href="/"
-        class="col-md-4 d-flex align-items-center justify-content-center mb-3 mb-md-0 me-md-auto link-body-emphasis text-decoration-none">
-        <svg class="bi me-2" width="40" height="32">
-          <use xlink:href="#bootstrap" />
-        </svg>
-      </a>
-      <ul class="nav col-md-4 justify-content-end">
-        <li class="nav-item"><a href="Landing.html" class="nav-link px-2 text-body-secondary">Home</a></li>
-        <li class="nav-item"><a href="About.html" class="nav-link px-2 text-body-secondary">About</a></li>
-      </ul>
-    </footer>
-  </div>
   <div class="d-flex" id="map"></div>
-  <script src="{{ url_for('static', filename='resources/scripts/mapScript.js') }}"></script>
-=======
   {% include 'footer.html' %}
->>>>>>> refs/remotes/origin/main
 </body>
 
 </html>
