+++ flaskr/templates/login.html
@@ -18,11 +18,11 @@
         <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
     
         <div class="form-floating">
-          <input type="email" class="form-control" id="floatingInput" name="email" placeholder="name@example.com">
+          <input type="email" class="form-control" id="floatingInput" name="email" placeholder="name@example.com" required>
           <label for="floatingInput">Email address</label>
         </div>
         <div class="form-floating">
-          <input type="password" class="form-control" id="floatingPassword" name="password" placeholder="Password">
+          <input type="password" class="form-control" id="floatingPassword" name="password" placeholder="Password" required>
           <label for="floatingPassword">Password</label>
         </div>
